Fix table creation for after Foreign key changes

+++ build_db/db_creation_queries/CREATE_CITY_TABLE.sql
@@ -2,7 +2,7 @@ CREATE TABLE City
 (
 	id MEDIUMINT(8) NOT NULL AUTO_INCREMENT,
 	name VARCHAR(38) NOT NULL,
-	country_id MEDIUMINT(8) NOT NULL,
+	country_id SMALLINT(5) NOT NULL,
 	PRIMARY KEY (id),
 	FOREIGN KEY (country_id) 
         	   REFERENCES  Country(id),
